Reorganized and fixed project structure
