+++ src/main/java/se/antons_skafferi/dataClass/DailyLunch.java
@@ -1,4 +1,4 @@
-package se.antons_skafferi.sqlDataClasses;
+package se.antons_skafferi.dataClass;
 
 import java.util.Date;
