+++ src/main/java/se/antons_skafferi/dataClass/Lunch.java
@@ -1,13 +1,10 @@
-package se.antons_skafferi.sqlDataClasses;
+package se.antons_skafferi.dataClass;
 
 import jakarta.persistence.Entity;
 import jakarta.persistence.GeneratedValue;
 import jakarta.persistence.GenerationType;
 import jakarta.persistence.Id;
 import jakarta.persistence.Table;
-import jakarta.persistence.Column;
-
-import java.util.Date;
 
 @Entity
 @Table(name = "lunch")
