+++ src/main/java/se/antons_skafferi/repository/BookingRepository.java
@@ -1,6 +1,7 @@
-package se.antons_skafferi.sqlDataClasses;
+package se.antons_skafferi.repository;
 
 import org.springframework.data.repository.CrudRepository;
+import se.antons_skafferi.dataClass.Bookings;
 
 import java.util.List;
