--- src/main/java/se/antons_skafferi/sqlDataClasses/AccessingDataMysqlApplication.java
@@ -1,13 +1,0 @@
-package se.antons_skafferi.sqlDataClasses;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class AccessingDataMysqlApplication {
-
-    public static void main(String[] args) {
-        SpringApplication.run(AccessingDataMysqlApplication.class, args);
-    }
-
-}
